src/backend/requirements.txt
pandas
fastapi
web3
pytest
uvicorn[standard]
psycopg2-binary